+++ tests/l2m2/client/test_base_llm_client.py
@@ -324,7 +324,7 @@ async def test_call_cerebras(mock_get_extra_message, mock_llm_post, llm_client):
     mock_get_extra_message.return_value = "extra message"
     mock_return_value = {"choices": [{"message": {"content": "response"}}]}
     mock_llm_post.return_value = mock_return_value
-    await _generic_test_call(llm_client, "cerebras", "llama-3.1-70b")
+    await _generic_test_call(llm_client, "cerebras", "llama-3.3-70b")
 
 
 @pytest.mark.asyncio